+++ Cargo.toml
@@ -22,5 +22,9 @@ codegen-units = 1
 [dependencies]
 clap = { version = "4", features = ["cargo", "wrap_help"] }
 enumflags2 = "0.7"
-light-curve-dmdt = {version = "0.6.1", path="light-curve-dmdt"}
 thiserror = "1"
+
+[dependencies.light-curve-dmdt]
+version = "0.6.1"
+path = "light-curve-dmdt"
+features = ["png"]
