+++ light-curve-dmdt/Cargo.toml
@@ -28,7 +28,7 @@ itertools = "0.10"
 libm = "0.2.1" # erf
 ndarray = "0.15.3"
 num-traits = "0.2.14"
-png = "0.17.1"
+png = { version = "0.17.1", optional = true }
 thiserror = "1"
 
 [dev-dependencies]
